Added the PieChart
In MainWindow:
Even though MainWindow is not the presenter, it should still interact with the DataView through the interface.  You are allowed to add things to the interface

Added to the interface (and corrected in MainWindow)
(1) bool ContextMenuEnabled
(2) void ResetFocus(int index) (moved from MainWindow to DataView)
(3) int SelectedIndex
(4) object SelectedItem
(5) int currentIndex

Errors:
(1) In MainWindow you did not use 'IDataView' for the view type, you used "DataView"
(2) You created a 'new' property 'presnter' (which was not part of the IDataView) instead of using the IDataView 'presenter' property that already existed.
(3) Used DataView.dataGrid.Items instead of DataView.DataSource
(4) You added a public property to DataView, but not to the interface (CurrentIndex)